Fix intersection with 0-area elements
